test(skill): split skill description suites by behavior
Split the oversized skill description and zauc skill-tool test suites into behavior-focused files while preserving the existing assertions, including the /customize-opencode native-skill regression coverage.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
